feat(LUN-25): Add userTable and sync with Clerk registered users
Added a new userTable to the database schema to store user information. Implemented functionality to sync this table with registered users from Clerk. This includes:  Defining the userTable schema. Creating a custom hook useAuthRegister to handle user registration and sync. Integrating the hook in the application to ensure users are added to the database upon registration with Clerk.