feat(LUN 26): Improve user appointment form and secure routes
Improved the user appointment form to capture the user's name, Clerk ID, and email using useUser(). The form now allows users to select a date and a restricted set of times, which are submitted as a timestamp. Added functionality to recognize and disable already occupied appointment slots on the frontend using the getAppointmentsByDateRange function. Ensured unique appointment timestamps on the backend by defining them as .unique in the schema. Updated the calendar UI design to align with the page's aesthetic. Secured all routes except for the landing page, sign-in, and sign-up pages in the middleware to ensure the user is always logged in when accessing the form. Replaced <a/> tags with Next.js Link components in intro.tsx for better navigation handling.